+++ Framework/Built_In_Automation/Sequential_Actions/action_declarations/info.py
@@ -41,6 +41,7 @@ action_support = (
     "sibling parameter",
     "parent parameter",
     "search element parameter",
+    "desired element parameter",
     "target parameter",
     "end parameter",
     "optional parameter",